Merge develop into main

+++ src/resource/api/RecordingResouce.py
@@ -1,6 +1,6 @@
 from flask import request
 from flask_restful import Resource
-from src.service import recording_service, user_service
+from src.service import recording_service, user_service, experiment_service
 from src.entity.RecordingState import RecordingState
 from flask import current_app
 from src.utils.authentication import authenticate
@@ -12,6 +12,11 @@ class RecordingResource(Resource):
 
     def __init__(self):
         self.socketio = current_app.socketio
+
+    @authenticate('api')
+    def get(self, payload: Payload):
+        assert payload.resource == 'user', f"Expected payload of resource: 'user' got {payload.resource}"
+        return recording_service.get_all(user=payload.id), 200
 
     @authenticate('api')
     def post(self, payload: Payload):
@@ -50,6 +55,14 @@ class RecordingResource(Resource):
 
         if recording.user != payload.id:
             return "You are not authorized to delete this recording", 401
+
+        experiment = experiment_service.find_by_recording(recording.id)
+
+        if experiment is not None:
+            return {
+                'error': 'An error occurred when stopping the recording',
+                'message': f'This recording is linked to <a href="/experiment/{experiment.id}">this</a> experiment you can\'t manually stop it!'
+            }, 400
 
         recording_service.delete(recording_id)
